database/requirements.txt
psycopg2-binary
pyaml
pytest
coverage